+++ packages/sdk/src/config/index.ts
@@ -1,11 +1,9 @@
-import { devConfig } from "./dev"
 import { prodConfig } from "./prod"
 import type { RaribleSdkConfig, RaribleSdkEnvironment } from "./domain"
 import { developmentConfig } from "./development"
 import { testnetConfig } from "./testnet"
 
 export const configsDictionary: Record<RaribleSdkEnvironment, RaribleSdkConfig> = {
-	dev: devConfig,
 	development: developmentConfig,
 	testnet: testnetConfig,
 	prod: prodConfig,