+++ packages/sdk/src/sdk-blockchains/solana/order.ts
@@ -12,7 +12,7 @@ import type {
 	PrepareOrderUpdateRequest,
 	PrepareOrderUpdateResponse,
 } from "../../types/order/common"
-import { OriginFeeSupport, PayoutsSupport } from "../../types/order/fill/domain"
+import { MaxFeesBasePointSupport, OriginFeeSupport, PayoutsSupport } from "../../types/order/fill/domain"
 import type { IApisSdk } from "../../domain"
 import type { CancelOrderRequest, ICancel } from "../../types/order/cancel/domain"
 import type {
@@ -73,6 +73,7 @@ export class SolanaOrder {
 		return {
 			originFeeSupport: OriginFeeSupport.NONE,
 			payoutsSupport: PayoutsSupport.NONE,
+			maxFeesBasePointSupport: MaxFeesBasePointSupport.IGNORED,
 			supportedCurrencies: getCurrencies(),
 			baseFee: await getAuctionHouseFee(auctionHouse, this.config?.auctionHouseMapping),
 			supportsExpirationDate: false,
@@ -114,6 +115,7 @@ export class SolanaOrder {
 		return {
 			originFeeSupport: OriginFeeSupport.NONE,
 			payoutsSupport: PayoutsSupport.NONE,
+			maxFeesBasePointSupport: MaxFeesBasePointSupport.IGNORED,
 			supportedCurrencies: getCurrencies(),
 			baseFee: await getAuctionHouseFee(auctionHouse, this.config?.auctionHouseMapping),
 			submit: updateAction,
@@ -164,6 +166,7 @@ export class SolanaOrder {
 			maxAmount: toBigNumber(item.supply),
 			originFeeSupport: OriginFeeSupport.NONE,
 			payoutsSupport: PayoutsSupport.NONE,
+			maxFeesBasePointSupport: MaxFeesBasePointSupport.IGNORED,
 			supportedCurrencies: getCurrencies(),
 			baseFee: 0,
 			getConvertableValue: this.getConvertableValue,
@@ -205,6 +208,7 @@ export class SolanaOrder {
 		return {
 			originFeeSupport: OriginFeeSupport.NONE,
 			payoutsSupport: PayoutsSupport.NONE,
+			maxFeesBasePointSupport: MaxFeesBasePointSupport.IGNORED,
 			supportedCurrencies: getCurrencies(),
 			baseFee: 0,
 			getConvertableValue: this.getConvertableValue,